refactor and added pizza id
