+++ src/app/pages/pizza/checkout/checkout.module.ts
@@ -2,19 +2,16 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CheckoutComponent } from './checkout.component';
 import { AntDesignCommonModule } from 'src/app/ant-design-common.module';
-import { ShareModule } from 'src/app/share/share.module';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { CompleteModule } from './complete/complete.module';
 import { StartModule } from './start/start.module';
+import { AppRoutingModule } from 'src/app/app-routing.module';
 
 @NgModule({
   declarations: [CheckoutComponent],
   imports: [
+    AppRoutingModule,
     AntDesignCommonModule,
-    ShareModule,
     CommonModule,
-    FormsModule,
-    ReactiveFormsModule,
     CompleteModule,
     StartModule
   ]
