+++ src/app/pages/pizza/checkout/complete/complete.component.html
@@ -1,1 +1,9 @@
-<p>complete works!</p>
+<nz-result
+    nzStatus="success"
+    nzTitle="Thank you for placing your pizza order with us!"
+    nzSubTitle="Order number id is: {{pizzaId}}" 
+>
+    <div nz-result-extra>
+        <button nz-button>View Order Status</button>
+    </div>
+</nz-result>
