+++ src/app/pages/pizza/checkout/start/start.component.css
@@ -1,0 +1,39 @@
+.center-text {
+    text-align: center;
+  }
+  .login-form {
+    max-width: 300px;
+  }
+  
+  .login-form-margin {
+    margin-bottom: 16px;
+  }
+  
+  .login-form-forgot {
+    float: right;
+  }
+  
+  .login-form-button {
+    width: 100%;
+  }
+  
+  .order-summary {
+    border-radius: 25px;
+    border: 2px solid green;
+    margin-left: 5px;
+    margin-right: 5px;
+    margin-bottom: 5px;
+    padding-left: 10px;
+    padding-bottom: 10px;
+  }
+  
+  .payment-info {
+    border-radius: 25px;
+    border: 2px solid green;
+    padding: 10px;
+    /* margin-left: 5px;
+    margin-right: 5px;
+    margin-bottom: 5px;
+    padding-left: 10px;
+    padding-bottom: 5px; */
+  }
