+++ src/app/pages/pizza/create/store/pizza.reducer.ts
@@ -8,7 +8,7 @@ export interface State {
 
 const initlaTasks: State = {
     pizza: {
-        'id': 0,
+        'id': (Math.floor(Math.random()*90000) + 10000),
         'size': [{'id': 0, 'label': 'SMALL', 'isSelected': true, 'value': 'sm'}, {'id': 1, 'label': 'MEDIUM', 'isSelected': false, 'value': 'md'}, {'id': 2, 'label': 'LARGE', 'isSelected': false, 'value': 'lg'}],
         'price': 10.99,
         'typeOfImage': {
